pyproject.toml
[tool.poetry]
name = "msds-scraper"
version = "1.0.0"
description = "Tool for scraping material safety datasheets"
authors = ["Unnatural Products Inc."]

[tool.poetry.scripts]
msds-scraper = "msds_scraper.cli:app"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
pandas = "^1.4.3"
typer = "^0.4.2"
openpyxl = "^3.0.10"
joblib = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-vcr = "^1.0.2"
python-magic = "^0.4.27"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
